requirements.txt
aiofiles==0.5.0
asyncpg>=0.20.0,<0.21.0
bcrypt>=3.1.0,<3.2.0
databases[postgresql]==0.2.6
fastapi>=0.55.0,<0.56.0
python-multipart==0.0.5
pyjwt>=1.7.0,<1.8.0
SQLAlchemy==1.3.12
uvicorn>=0.11.0,<0.12.0
gunicorn

# dev
pytest==5.3.2
requests==2.22.0